\name{fbpack-package}
\alias{fbpack-package}
\alias{fbpack}
\docType{package}
\title{Facebook crawler package}
\description{
This package provides user-friendly access to Facebook from R only needing a user name and a password. This way one can look at the site through R code and see what one would see in a browser.
}
\details{
\tabular{ll}{
Package: \tab fbpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-19\cr
License: \tab GLP-3\cr
Depends: \tab RSelenium, RCurl, RJSONIO, XML\cr
}}

\author{
Andras Tajti

Maintainer: Andras Tajti <atajti@gmail.com>

}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[RSelenium:RSelenium-package]{RSelenium}} ~~
}
\examples{
## NOT RUN
# search_fb("anyone", "your_id", "your_password")
## END NOT RUN
}
